Adds a more useful error message.
Extending the Error Message when we set a version schema which isn't allowed. As an new user which tries your build management tool for the first time, I run into this as first issue, and the documentation didn't tell me anything about this configuration property and what the version schema has to do with my build. 

This quick-win, of self documentation an throwing an Error should be helpful for the next new user of habitus.
Sorry I have no experience with golang, I just hope that "\n works."